# api/openapi.yaml
openapi: 3.0.3
info:
  title: asyou Management API
  version: v1
servers:
  - url: /api/v1
paths:
...
  /nodes:
    post:
      summary: Create a frps node
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeCreate'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
    get:
      summary: List nodes
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Node'
  /nodes/{id}:
    get:
      summary: Get node
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
    put:
      summary: Update node (partial)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeUpdate'
      responses:
        '200':
          description: updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
    delete:
      summary: Delete node
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '204':
          description: deleted
  /nodes/{id}/heartbeat:
    post:
      summary: Node heartbeat (auth by node token header)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeHeartbeat'
      responses:
        '204':
          description: accepted
...
  /proxies:
    post:
      summary: Create proxy
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyCreate'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proxy'
    get:
      summary: List proxies for user
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Proxy'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
...
    NodeCreate:
      type: object
      required: [name, host]
      properties:
        name:
          type: string
        host:
          type: string
        api_port:
          type: integer
        bind_port:
          type: integer
        tls_enabled:
          type: boolean
        auth_token:
          type: string
    NodeUpdate:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        api_port:
          type: integer
        bind_port:
          type: integer
        tls_enabled:
          type: boolean
        auth_token:
          type: string
    Node:
      allOf:
        - $ref: '#/components/schemas/NodeCreate'
        - type: object
          properties:
            id:
              type: integer
            last_heartbeat:
              type: string
              format: date-time
    NodeHeartbeat:
      type: object
      properties:
        uptime:
          type: integer
        connections:
          type: integer
        load:
          type: number
    ProxyCreate:
      type: object
      required: [name, type, local_port]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [tcp, http, https, udp, stcp, xtcp, tcpmux]
        local_ip:
          type: string
          default: '127.0.0.1'
        local_port:
          type: integer
        remote_port:
          type: integer
        subdomain:
          type: string
        custom_domains:
          type: array
          items:
            type: string
        node_id:
          type: integer
    Proxy:
      allOf:
        - $ref: '#/components/schemas/ProxyCreate'
        - type: object
          properties:
            id:
              type: integer
            user_id:
              type: integer
            status:
              type: string
            annotations:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...